backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["\"Steve Bickely\""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
djangorestframework = "3.14.0"
django-cors-headers = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
pyjwt = "2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
